[iOS] Refactor: 뷰모델 객체가 데이터소스를 채택하는 형식으로 구조를 바꿈 	이유: 뷰모델이란 뷰를 그리기 위한 데이터를 가지고 있고, 테이블 뷰 데이터소스는 	      테이블 뷰를 그리기 위한 데이터를 갖고 있어야 하므로 서로 똑같은 일을 하고 있는 것이다. 		따라서 뷰모델 객체가 테이블 뷰 데이터소스의 역할도 하는 것이 적절하기 때문에 채택하도록 하였다.
